+++ CHANGELOG.md
@@ -56,6 +56,8 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 ## Wazuh v4.5.0 - OpenSearch Dashboards 2.6.0 - Revision 01
 
+### Added
+
 - Support for Wazuh 4.5.0
 
 ## Wazuh v4.4.4 - OpenSearch Dashboards 2.6.0 - Revision 01