+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the Wazuh app project will be documented in this file.
 
-## Wazuh v4.4.5 - OpenSearch Dashboards 2.6.0 - Revision 00
+## Wazuh v4.5.1 - OpenSearch Dashboards 2.6.0 - Revision 01
 
 ### Added
 
@@ -20,6 +20,12 @@ All notable changes to the Wazuh app project will be documented in this file.
 ### Changed
 
 - Changed windows agent service name in the deploy agent wizard [#5538](https://github.com/wazuh/wazuh-kibana-app/pull/5538)
+
+## Wazuh v4.5.0 - OpenSearch Dashboards 2.6.0 - Revision 01
+
+### Added
+
+- Support for Wazuh 4.5.0
 
 ## Wazuh v4.4.4 - OpenSearch Dashboards 2.6.0 - Revision 01